--- srcs/parsing/ft_free.c
@@ -1,14 +1,0 @@
-#include "../../include/cub3D.h"
-
-void ft_free(char **str)
-{
-	int i;
-
-	i = 0;
-	while (str[i])
-	{
-		free(str[i]);
-		i++;
-	}
-	free(str);
-}
