+++ srcs/parsing/get_map.c
@@ -1,6 +1,6 @@
 #include "../../include/cub3D.h"
 
-void	reverse_map(t_game_info *game)
+static void	reverse_map(t_game_info *game)
 {
 	int		i;
 	int		j;
@@ -18,7 +18,6 @@ void	reverse_map(t_game_info *game)
 	}
 	game->ray.player.y = game->map_height - game->ray.player.y;
 }
-
 
 void	get_map(char *map_file, t_game_info *game)
 {
