+++ srcs/raycasting/choose_texture.c
@@ -2,7 +2,7 @@
 
 void choose_texture(t_game_info *game, t_raycast *ray)
 {
-	if (ray->side == DECIDE_N_OR_S && ray->ray_dir.y < 0)		// x > 0
+	if (ray->side == DECIDE_N_OR_S && ray->ray_dir.y < 0)
 		ray->wall_type = &game->so_texture;
 	else if (ray->side == DECIDE_N_OR_S && ray->ray_dir.y > 0)
 		ray->wall_type = &game->no_texture;
