+++ test_maps/dir_test.cub
@@ -8,7 +8,7 @@ C 153,204,255
 
 1111111111111111111
 1010000000000000001
-11N1000000000000001
+11S0000000000000001
 1010000000000000001
 1000000000000000001
 1000000000000000001
